+++ app/src/js/reducers/index.js
@@ -9,6 +9,7 @@ import config from './utils/config';
 import dist from './dist';
 import datepicker from './datepicker';
 import granules from './granules';
+import granulesExecutions from './granules-executions';
 import stats from './stats';
 import pdrs from './pdrs';
 import providers from './providers';
@@ -39,6 +40,7 @@ export const reducers = {
   datepicker,
   cumulusInstance,
   granules,
+  granulesExecutions,
   stats,
   timer,
   sidebar,