+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- **CUMULUS-2462**
+  - Reingest a granule from granule details page is updated to select a workflow execution arn for reingest.
+  - Batch Reingest and Bulk Reingest Actions are updated to select the workflowName for reingest.
+
 ## [v7.0.0]
 
 ## Breaking Changes